(#3344) Adds Param Block Handling to chocolateyScriptRunner
Adds a basic function to the helper functions that parses a script file for parameters, and compares that list to available Package Parameters.

If there are matches, it returns a hashtable with matching names and their values.

This can then be used in the chocolateyScriptRunner to pass parameters into the param block.

The methods used all work back to PowerShell v3, though without any plan for backwards compatibility (e.g. considerations for using an extension package/module) it's hard to see this being that popular.