() Create License Validation Method That Accepts a Path
There's not currently a way to validate a license that isn't in place at the correct location.

This change splits the existing Validation method into two parts.
A method that accepts a path, and a method that
needs no parameter (which calls the path method with the previously used
path).